Only allows owners to delete scavenger hunts

+++ db/sql/hunts/remove.sql
@@ -1,2 +1,3 @@
 DELETE FROM hunts
-WHERE hunt_id = $1;
+WHERE hunt_id = ${hunt_id}
+AND owner_id = ${user_id};
